requirements.txt
alembic==1.14.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
# greenlet==3.1.1  # Python 3.14 ile uyumsuz, opsiyonel
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.1
numpy>=2.1.0
scipy>=1.14.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
